Add simpler mesh2d example
