+++ Cargo.toml
@@ -121,6 +121,10 @@ name = "mesh2d"
 path = "examples/2d/mesh2d.rs"
 
 [[example]]
+name = "mesh2d_manual"
+path = "examples/2d/mesh2d_manual.rs"
+
+[[example]]
 name = "rect"
 path = "examples/2d/rect.rs"
